remote_agents/sports_results_agent/pyproject.toml
[project]
name = "sports-results-agent"
version = "0.1.0"
description = "Multi-agent system with Azure AI Agents integration for routing and task delegation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.2.1",
    "a2a-sdk==0.3.10",
    "openai-agents>=0.1.0",
]